Fix the image add to the status bar
The Gallery is killed by system in background due to some reason.
The Gallery restore last states after restart it.
But the mediaset is loading, the reload function return immediately,
and the data is empty,the Photopage is finished when the data is empty.
So the UI is wrong.

If the mediaset is loading, waiting for it. After loading successfully,
display the right photo.

Change-Id: I384231bd546c5bb8984757d1b3e51c10220c6ca6