Fix picture is different when changes font size
when change the system font size, photopage destroy. when re-start it,
the photopage restore from stack. But the current photo index is not saved.
So photopage display the original photo not the current photo.

In onSaveState the current photo index is saved. In onCreate that current
index is restored if the photopage restore from stack.

Change-Id: Iedcc0809eee1920d4bd2466bfc8ea1201c375b26