Fix the picture display error after doing rotate operation.
After rotating the image, use previous canvas to draw the new content,
but the canvas is not erased the previous content before using to draw.
So if the format of the image is png which alpha is set,
the previous content can be seen by the transparent part.

Before drawing the new content, the canvas erase the previous content to
make the canvas clean.

Change-Id: I4831ed052993955b7296e589760d657eb40a0121